refactor: move site auto create logic to upper layer
