+++ src/constants.js
@@ -1,1 +1,7 @@
 export const THEME_COLOR = '#C3AAD1';
+export const SITE_CONFIG = {
+  config: {display: '設定', value: 'config'},
+  landing: {display: '首頁設定', value: 'landing'},
+  product_category: {display: '產品分類', value: 'product_category'},
+  article_category: {display: '文章分類', value: 'article_category'},
+};
